JSON Schema: Cleanup and documentation (#45)
Cleaned up the JSON Schema package documentation a bit, as well as some
of the code. I changed one of the tests and the ErrNotImplemented error
message. Pacakge impors was rearanged by goimports.

I used the Go standard library as a guide line for the docs, including
the use of full sentences with a trailing period in comments and short,
lower case error messages.