Refactor of the resource.Lookup to be part of the query package
Having pagination and projection in the resource package (Lookup type)
and the query in the schema/query package feels odd. If we want the
schema and the query package to be really idependant, all the tools to
perform a query should be in the query package.

This change moves the resource.Lookup type into the query.Query. The
previous query.Query type becomes query.Predicate and is now a field of
the query.Query type. The former resource.Lookup selector field becomes
the Projection field of the query.Query type and the pagination (offset
and limit) are now embedded into the query instead of being passed as
separate parameters to the Storer.

This change breaks compatibility with Storers. They all now have to
replace resource.Lookup, offset and limit parameters by this new
query.Query type.

One cool thing about this, is that the previously private connection
validator to handle connection in the resource package is now a public
validator in the schema package. The query.Projection Eval method is now
able to handle connections and references without depending on the
resource package. This will helps people wanting to use schema and
schema/query independantly from the rest of the project. This does also
make the code cleaner and easier to understand.

I am still not happy with the way we handle sub-requests for connection
and reference resolution. The coalecing code is quite messy, and there
is no good test coverage. The test quality for the projection type is not
great easier. Next commit will refactor the sub-request code and improve
tests.

Issue #112