Remove query predicate simplifier
When a parsed predicate does only contain a single $and, we use to
simplify the query by removing the wrapping And type.

This creates issues when the end contains tests on the same field like
for instance:

   {$and: [{foo: {$ne: "bar"}}, {foo: {$ne: "baz"}}]}

Simplified this would result in:

   {foo: {$ne: "bar"}, foo: {$ne: "baz"}}

Which as JSON would fold the keys and only keep the last condition.