Fix demo
The demo uses a dependency referencing a field from a parent schema.
This does not work with the current implementation, and I don't recall
it working in the past either. We might want to make such dependency
work, but it might require some major changes.

I thus removed the sub-schema and made the demo more realistic.

Fix #118