Array support for index lookup (breaking change)
This commit adds support for looking up field definitions inside
schema.Array instances. To achieve this, Array.ValuesValidator has been
deprecated in favour of a new Array.Values property that holds a full
schema.Field.

As a result, it should be possible to filter on schema.Array index value
in backends that supports it.